+++ apps/app01/web.py
@@ -17,6 +17,8 @@ def index():
   env_app_name = os.environ.get('APP_NAME')
   cd_stage =  os.environ.get('CLOUD_DEPLOY_TARGET')
   ric_msg = os.environ.get('RICCARDO_MESSAGE')
+  # moving vars to COMMON from /components/
+  fav_color_common = os.environ.get('FAVORITE_COLOR_COMMON')
   cloud_deploy_target_common = os.environ.get('CLOUD_DEPLOY_TARGET_COMMON')
 
   print("[ricc][web.py] INDEX: the super-duper vanilla server in python to say HelloWorld - v{}!\n".format(version))
@@ -25,16 +27,17 @@ def index():
         Hello world from Skaffold in python! This is a demonstrative app to demonstrate CI/CD with Cloud Deploy and Cloud Build<br/>
 
         I read version from file and this ./VERSION file is actually read by the build pipeline
-        into the Cloud Deploy release name - wOOOt!<br/>
+        into the Cloud Deploy release name - wOOOt!<br/><br/>
 
         FAVORITE_COLOR={fav_color}<br/>
         CD_TARGET={cd_stage} <br/>
-        CLOUD_DEPLOY_TARGET_COMMON={cloud_deploy_target_common} <br/>
+        CLOUD_DEPLOY_TARGET_COMMON={cloud_deploy_target_common} <br/><br/>
 
         APP_NAME={env_app_name} <br/>
-        RICCARDO_MESSAGE={ric_msg}
+        RICCARDO_MESSAGE={ric_msg}<br/><br/>
 
-        Favorite Color from v14: <b style='background-color:#{fav_color};' >#{fav_color}</b><br/>
+        Favorite Color COMMON from v1.35: <b style='background-color:#{fav_color_common};' >#{fav_color_common}</b><br/>
+        Favorite Color from v1.4: <b style='background-color:#{fav_color};' >#{fav_color}</b><br/>
 
         <hr/>
           <center>
@@ -43,6 +46,7 @@ def index():
   """.format(
     ver=version,
     fav_color=fav_color,
+    fav_color_common=fav_color_common,
     env_app_name=env_app_name,
     cd_stage=cd_stage,
     cloud_deploy_target_common=cloud_deploy_target_common,